Remove Pattern class for now
It didn't do anything besides validate stitch counts. Not sure how this
will work since the pattern class lost all information about repeats,
etc. that we want to keep in the output, so we might need to replace it
with something else that doesn't evaluate the AST all the way.