Remove RowRepeatExpr case from _merge_across
Row repeats have been already completely unrolled by flatten, so this
case is only ever run when there's a PatternExpr (which is a subclass of
RowRepeatExpr with times = 1). So I don't think the lcm/gcd stuff
actually gets a chance to do anything? We'd have to tell flatten to not
always unroll.